Some PEP8 love and handle unicode values.
Noticed that not all values are returned as 'plain' strings and
sometimes you can get a unicoded value, which screws up some of the
string interpolation for generating reports. I have changed the report
to decode to UTF-8 automatically for the QA Contact and Description
fields. I have also done some fixes to be more PEP8-friendly.